[#8] feat: PR 리뷰 반영
- int 오버플로우 방지

+++ src/main/java/com/kbo/seatblock/controller/request/SeatBlockRequest.java
@@ -1,6 +1,6 @@
 package com.kbo.seatblock.controller.request;
 
 public record SeatBlockRequest(
-	String name, long seatCount, long stadiumId
+	String name, int seatCount, long stadiumId
 ) {
 }
