+++ src/test/java/com/kbo/fixture/SeatBlockFixture.java
@@ -4,7 +4,7 @@ import com.kbo.seatblock.entity.SeatBlock;
 import com.kbo.stadium.entity.Stadium;
 
 public class SeatBlockFixture {
-	public static SeatBlock get(String name, long seatCount, Stadium stadium) {
+	public static SeatBlock get(String name, int seatCount, Stadium stadium) {
 		return new SeatBlock(name, seatCount, stadium);
 	}
 }
